# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Hana Omni-Connect API
  version: 0.1.0-feature
  summary: B2B Korea market data and watchlist alert API
  description: >
    Hana Omni-Connect API provides Korea stock information, foreign ownership
    snapshots, and news/disclosure alert events for overseas exchange and broker
    partners. It does not provide order, execution, settlement, FX trading, or
    omnibus account processing. Every response includes
    X-HANA-OMNI-CONNECT-CORRELATION-ID for support and incident tracing.
servers:
  - url: https://api.hana-omni-connect.example.com
    description: Production placeholder
  - url: http://localhost:8080
    description: Local development
security:
  - partnerApiKey: []
...
paths:
  /api/v1/portal/auth/sign-up:
    post:
      tags: [Portal]
      security: []
      summary: Create a portal member account with server-side password confirmation.
      responses:
        "200": { description: Signed-in portal session. }
        "400": { $ref: "#/components/responses/ValidationError" }
  /api/v1/portal/auth/login:
    post:
      tags: [Portal]
      security: []
      summary: Sign in to the portal.
      responses:
        "200": { description: Portal session and password-change requirement. }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/alerts/watchlists/{partnerId}:
    get:
      tags:
        - Alerts
      operationId: getPartnerWatchlist
      summary: Get a partner watchlist used by scheduled alert collection.
      parameters:
        - $ref: "#/components/parameters/PartnerId"
      responses:
        "200":
          description: Stored partner watchlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerWatchlistApiResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    put:
      tags:
        - Alerts
      operationId: replacePartnerWatchlist
      summary: Replace a partner watchlist used by scheduled alert collection.
      description: >
        Replaces the full list for one partner. Duplicate stock codes are
        removed while preserving request order. Empty lists clear the partner
        watchlist.
      parameters:
        - $ref: "#/components/parameters/PartnerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerWatchlistUpdateRequest"
      responses:
        "200":
          description: Stored partner watchlist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerWatchlistApiResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/PartnerAccessDenied"
        "404":
          $ref: "#/components/responses/StockNotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/alerts/events:
    post:
      tags:
        - Alerts
      operationId: publishAlertEvent
      summary: Publish an already analyzed alert event to partner and stock topics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertPublishRequest"
      responses:
        "200":
          description: Published alert event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertEventApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/PartnerAccessDenied"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/alerts/analyze-and-publish:
    post:
      tags:
        - Alerts
      operationId: analyzeAndPublishAlert
      summary: Analyze one news or disclosure item with Hannah-Montana-AI and publish it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertAnalysisPublishRequest"
      responses:
        "200":
          description: Published alert event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertEventApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/PartnerAccessDenied"
        "422":
          description: Hannah-Montana-AI did not match the item to a supported stock.
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/alerts/collect-and-publish:
    post:
      tags:
        - Alerts
      operationId: collectAnalyzeAndPublishAlerts
      summary: Collect provider news and disclosures for stock codes, analyze them, and publish alerts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertCollectPublishRequest"
      responses:
        "200":
          description: Collection and publishing result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertCollectPublishApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/PartnerAccessDenied"
        "404":
          description: None of the requested stock codes are supported.
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/korean-financial-terms/explain:
    post:
      tags:
        - Korean Financial Terms
      operationId: explainKoreanFinancialTerm
      summary: Explain a clicked Korean financial term and record usage analytics.
      description: >
        Returns an English explanation for Korean market-specific financial
        slang or terminology from the Hannah-Montana-AI dictionary/RAG engine.
        Every request records anonymized click analytics so high-frequency terms
        can be cached and low-confidence new terms can be reviewed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KoreanFinancialTermExplainRequest"
      responses:
        "200":
          description: Korean financial term explanation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoreanFinancialTermExplanationApiResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/korean-financial-terms/stats:
    get:
      tags:
        - Korean Financial Terms
      operationId: listKoreanFinancialTermStats
      summary: List clicked Korean financial term analytics.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        "200":
          description: Korean financial term click statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoreanFinancialTermClickStatListApiResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    partnerApiKey:
      type: apiKey
      in: header
      name: X-HANA-OMNI-CONNECT-API-KEY
      description: Partner API key. The server stores only SHA-256 hashes and can bind keys to partner IDs.
    portalBearer:
      type: http
      scheme: bearer
      bearerFormat: HMAC-SHA256
      description: Short-lived portal session token. Password changes revoke earlier token versions.
  parameters:
...
    PartnerId:
      name: partnerId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 80
        pattern: "^[A-Za-z0-9._:-]+$"
      description: Partner identifier issued to an overseas exchange or broker.
  responses:
    Unauthorized:
      description: Missing or invalid partner API key.
    ValidationError:
      description: Request validation failed.
    StockNotFound:
      description: Supported stock master row was not found.
    PartnerAccessDenied:
      description: Authenticated partner cannot access another partner's alert resource.
    ServiceUnavailable:
      description: Server security or provider configuration is incomplete.
  schemas:
...
    ApiResponseBase:
      type: object
      required:
        - success
        - status
        - code
        - message
        - timestamp
      properties:
        success:
          type: boolean
        status:
          type: integer
        code:
          type: string
          example: COMMON_000
        message:
          type: string
          example: OK
        errors:
          type:
            - array
            - "null"
          items:
            type: object
        timestamp:
          type: string
          format: date-time
    KoreanFinancialTermExplainRequest:
      type: object
      required:
        - term
      properties:
        term:
          type: string
          minLength: 1
          maxLength: 80
          example: 개미
        locale:
          type: string
          enum: [en]
          default: en
        sourceType:
          type: string
          enum: [NEWS, DISCLOSURE]
          default: NEWS
        title:
          type: string
          maxLength: 300
        context:
          type: string
          maxLength: 4000
        stockCode:
          type: string
          pattern: "^\\d{6}$"
        stockName:
          type: string
          maxLength: 80
        articleId:
          type: string
          maxLength: 120
        articleUrl:
          type: string
          maxLength: 1000
        userKey:
          type: string
          maxLength: 160
          description: Partner user identifier. The server stores only a salted hash.
        sessionKey:
          type: string
          maxLength: 160
          description: Partner session identifier. The server stores only a salted hash.
...
    KoreanFinancialTermExplanationApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          required:
            - data
          properties:
            data:
              $ref: "#/components/schemas/KoreanFinancialTermExplanation"
...
    KoreanFinancialTermClickStatListApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/KoreanFinancialTermClickStat"
...
    PartnerWatchlistApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          required:
            - data
          properties:
            data:
              $ref: "#/components/schemas/PartnerWatchlistResponse"
...
    AlertEventApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          required:
            - data
          properties:
            data:
              $ref: "#/components/schemas/AlertEvent"
    AlertCollectPublishApiResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          required:
            - data
          properties:
            data:
              $ref: "#/components/schemas/AlertCollectPublishResponse"
...
    AlertPublishRequest:
      type: object
      required:
        - partnerId
        - stockCode
        - stockName
        - sourceType
        - originalTitle
        - translatedTitle
        - summary
        - originalUrl
        - publishedAt
        - sentiment
        - importance
      properties:
        partnerId:
          type: string
          maxLength: 80
        stockCode:
          type: string
          pattern: "^\\d{6}$"
        stockName:
          type: string
          maxLength: 80
        sourceType:
          $ref: "#/components/schemas/SourceType"
        originalTitle:
          type: string
          maxLength: 300
        translatedTitle:
          type: string
          maxLength: 300
        summary:
          type: string
          maxLength: 1000
        summaryLines:
          anyOf:
            - $ref: "#/components/schemas/AlertSummaryLines"
            - type: "null"
        translatedSummary:
          type:
            - string
            - "null"
          maxLength: 2000
        originalContent:
          type:
            - string
            - "null"
          maxLength: 1000000
          description: 문단과 줄바꿈을 보존한 뉴스 또는 공시 원문이다.
        translatedContent:
          type:
            - string
            - "null"
          maxLength: 1000000
        imageUrls:
          type: array
          items:
            type: string
            maxLength: 1000
        contentAvailability:
          type:
            - string
            - "null"
          enum:
            - FULL_TEXT
            - ORIGINAL_TEXT_ONLY
            - SUMMARY_ONLY
            - null
          description: 원문과 번역 전문의 실제 저장 상태다.
        originalUrl:
          type: string
          maxLength: 500
        publishedAt:
          type: string
          format: date-time
        eventTags:
          type: array
          items:
            type: string
        sentiment:
          $ref: "#/components/schemas/Sentiment"
        importance:
          $ref: "#/components/schemas/Importance"
        marketImpactImportance:
          description: Hana Montana AI(KF-DeBERTa + K-FNSPID)가 예측한 사후 가격반응 등급이다. 의미 중요도와 독립된 보조 신호이며 세 시장영향 필드는 함께 제공하거나 모두 생략한다.
          anyOf:
            - $ref: "#/components/schemas/Importance"
            - type: "null"
        marketImpactScore:
          type:
            - number
            - "null"
          minimum: 0
          maximum: 1
          description: K-FNSPID v4의 시장반응 순서형 점수다. 투자수익이나 인과효과를 의미하지 않는다.
        marketImpactConfidence:
          type:
            - number
            - "null"
          minimum: 0
          maximum: 1
          description: 파일 기반 시장영향 모델의 예측 confidence다. 의미 중요도 confidence와 합산하지 않는다.
        relatedStocks:
          type: array
          items:
            type: string
            pattern: "^\\d{6}$"
        holderTarget:
          type: boolean
        watchlistTarget:
          type: boolean
        glossaryTerms:
          type: array
          description: Financial glossary terms normalized by Hannah-Montana-AI for translation quality review.
          items:
            $ref: "#/components/schemas/AlertGlossaryTerm"
        translationQualityFlags:
          type: array
          description: Translation quality and review flags propagated from Hannah-Montana-AI.
          items:
            type: string
          example:
            - FINANCIAL_GLOSSARY_APPLIED
        translationProvider:
          type: string
          description: Active translation provider or source-language fallback marker.
          example: local-open-source-qwen3-translation
        translationModelVersion:
          type: string
          description: Translation model name used by the provider when available.
          example: local-llm:Qwen3-4B-GGUF-Q4
        translationStatus:
          type: string
          enum:
            - TRANSLATED
            - PARTIAL_SOURCE_LANGUAGE_FALLBACK
            - SOURCE_LANGUAGE_FALLBACK
          description: Whether translated fields came from Hannah Qwen3 or were returned as source-language fallback.
        duplicateKey:
          type:
            - string
            - "null"
          maxLength: 128
          description: AI normalized duplicate key for downstream idempotency.
        clusterKey:
          type:
            - string
            - "null"
          maxLength: 128
          description: 동일 사건의 반복 기사와 공시를 묶는 시간창 기반 cluster key다.
        modelVersion:
          type:
            - string
            - "null"
          maxLength: 240
          description: Hana Montana AI(KF-DeBERTa + K-FNSPID)의 event, sentiment, semantic materiality, market-impact 복합 출처다. OmniConnect는 파싱하거나 변경하지 않고 전파한다.
        eventConfidence:
          type:
            - number
            - "null"
          minimum: 0
          maximum: 1
          description: Event tag confidence propagated from Hannah-Montana-AI for observation and UI display only. It is not an automatic blocking decision.
        sentimentConfidence:
          type:
            - number
            - "null"
          minimum: 0
          maximum: 1
          description: Sentiment confidence propagated from Hannah-Montana-AI for observation and UI display only. It is not an automatic blocking decision.
        importanceConfidence:
          type:
            - number
            - "null"
          minimum: 0
          maximum: 1
          description: Importance confidence propagated from Hannah-Montana-AI for observation and UI display only. It is not an automatic blocking decision.
        stockMatchConfidence:
          type:
            - number
            - "null"
          minimum: 0
          maximum: 1
          description: Stock mapping confidence propagated from Hannah-Montana-AI for observation and UI display only. It is not an automatic blocking decision.
    AlertSummaryLines:
      type: object
      required:
        - what
        - why
        - impact
      properties:
        what:
          type: string
          description: 발생한 사실과 주체를 근거 범위에서 요약한다.
        why:
          type: string
          description: 기사 또는 공시에 명시된 배경을 요약한다.
        impact:
          type: string
          description: 의미 중요도와 감성에 따른 확인 사항을 요약한다.
...
    AlertAnalysisPublishRequest:
      type: object
      required:
        - partnerId
        - sourceType
        - title
        - originalUrl
        - publishedAt
      properties:
        partnerId:
          type: string
          maxLength: 80
        sourceType:
          $ref: "#/components/schemas/SourceType"
        title:
          type: string
          maxLength: 300
        snippet:
          type: string
          maxLength: 1000
        originalUrl:
          type: string
          maxLength: 500
        publishedAt:
          type: string
          format: date-time
        stockUniverse:
          type: array
          items:
            $ref: "#/components/schemas/StockCandidate"
...
    AlertCollectPublishRequest:
      type: object
      required:
        - partnerId
        - stockCodes
      properties:
        partnerId:
          type: string
          maxLength: 80
        stockCodes:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: string
            pattern: "^\\d{6}$"
        newsDisplay:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        disclosureLookbackDays:
          type: integer
          minimum: 1
          maximum: 30
          default: 7
    PartnerWatchlistUpdateRequest:
      type: object
      required:
        - stockCodes
      properties:
        stockCodes:
          type: array
          maxItems: 100
          items:
            type: string
            pattern: "^\\d{6}$"
    PartnerWatchlistResponse:
      type: object
      required:
        - partnerId
        - stockCodes
      properties:
        partnerId:
          type: string
          maxLength: 80
        stockCodes:
          type: array
          items:
            type: string
            pattern: "^\\d{6}$"
...
    AlertEvent:
      allOf:
        - $ref: "#/components/schemas/AlertPublishRequest"
        - type: object
          required:
            - alertId
            - createdAt
          properties:
            alertId:
              type: string
            createdAt:
              type: string
              format: date-time
    SourceType:
      type: string
      enum:
        - NEWS
        - DISCLOSURE
    Sentiment:
      type: string
      enum:
        - POSITIVE
        - NEUTRAL
        - NEGATIVE
    Importance:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
        - CRITICAL